+++ drivers/net/wireless/ti/wlcore/main.c
@@ -3234,7 +3234,7 @@ static uintptr_t wl1271_op_prepare_multicast(struct ieee80211_hw *hw,
 		}
 	}
 
-	return (u64)(uintptr_t)fp;
+	return (uintptr_t)fp;
 }
 
 #define WL1271_SUPPORTED_FILTERS (FIF_ALLMULTI | \
@@ -3248,7 +3248,7 @@ static void wl1271_op_configure_filter(struct ieee80211_hw *hw,
 				       unsigned int *total,
 				       uintptr_t multicast)
 {
-	struct wl1271_filter_params *fp = (void *)(unsigned long)multicast;
+	struct wl1271_filter_params *fp = (void *)multicast;
 	struct wl1271 *wl = hw->priv;
 	struct wl12xx_vif *wlvif;
